requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
azure-identity
azure.monitor.query
semantic-kernel